goalkeeper_back/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib
python-dotenv
httpx
pydantic
python-multipart
pyjwt
google-auth
requests